+++ packages/vrender-components/src/axis/tick-data/continuous.ts
@@ -4,8 +4,8 @@ import { isContinuous } from '@visactor/vscale';
 import { isFunction, isValid, last } from '@visactor/vutils';
 import type { ICartesianTickDataOpt, ILabelItem, ITickData, ITickDataOpt } from '../type';
 // eslint-disable-next-line no-duplicate-imports
-import { convertDomainToTickData, getCartesianLabelBounds, hasOverlap, intersect } from './util';
-
+import { convertDomainToTickData, getCartesianLabelBounds } from './util';
+import { textIntersect as intersect, hasOverlap } from '../util';
 function getScaleTicks(
   op: ITickDataOpt,
   scale: ContinuousScale,
@@ -38,6 +38,34 @@ function getScaleTicks(
   return scaleTicks;
 }
 
+function forceItemVisible(
+  sourceItem: ILabelItem<number>,
+  items: ILabelItem<number>[],
+  check: boolean,
+  comparator: any,
+  inverse = false
+) {
+  if (check && !items.includes(sourceItem)) {
+    let remainLength = items.length;
+    if (remainLength > 1) {
+      if (inverse) {
+        items.push(sourceItem);
+      } else {
+        items.unshift(sourceItem);
+      }
+      for (let i = 0; i < remainLength; i++) {
+        const index = inverse ? remainLength - 1 - i : i;
+        if (comparator(items[index])) {
+          items.splice(index, 1);
+          i--;
+          remainLength--;
+        } else {
+          break;
+        }
+      }
+    }
+  }
+}
 /** 连续轴默认 tick 数量 */
 export const DEFAULT_CONTINUOUS_TICK_COUNT = 5;
 /**
@@ -97,7 +125,20 @@ export const continuousTicks = (scale: ContinuousScale, op: ITickDataOpt): ITick
     });
   }
 
-  if (op.sampling) {
+  const domain = scale.domain();
+
+  if (op.labelFirstVisible && domain[0] !== scaleTicks[0] && !scaleTicks.includes(domain[0])) {
+    scaleTicks.unshift(domain[0]);
+  }
+
+  if (
+    op.labelLastVisible &&
+    domain[domain.length - 1] !== scaleTicks[scaleTicks.length - 1] &&
+    !scaleTicks.includes(domain[domain.length - 1])
+  ) {
+    scaleTicks.push(domain[domain.length - 1]);
+  }
+  if (op.sampling && scaleTicks.length > 1) {
     // 判断重叠
     if (op.coordinateType === 'cartesian' || (op.coordinateType === 'polar' && op.axisOrientType === 'radius')) {
       const { labelGap = 4, labelFlush } = op as ICartesianTickDataOpt;
@@ -108,10 +149,40 @@ export const continuousTicks = (scale: ContinuousScale, op: ITickDataOpt): ITick
             value: scaleTicks[i]
           } as ILabelItem<number>)
       );
+      const source = [...items];
+      const firstSourceItem = source[0];
+      const lastSourceItem = last(source);
+
       const samplingMethod = breakData && breakData() ? methods.greedy : methods.parity; // 由于轴截断后刻度会存在不均匀的情况，所以不能使用 parity 算法
-      while (items.length >= 3 && hasOverlap(items, labelGap)) {
+      while (items.length >= 3 && hasOverlap(items as any, labelGap)) {
         items = samplingMethod(items, labelGap);
       }
+
+      const shouldCheck = (length: number, visibility: boolean) => length < 3 || visibility;
+
+      const checkFirst = shouldCheck(items.length, op.labelFirstVisible);
+      let checkLast = shouldCheck(items.length, op.labelLastVisible);
+
+      if (intersect(firstSourceItem as any, lastSourceItem as any, labelGap)) {
+        if (items.includes(lastSourceItem) && items.length > 1) {
+          items.splice(items.indexOf(lastSourceItem), 1);
+          checkLast = false;
+        }
+      }
+
+      forceItemVisible(firstSourceItem, items, checkFirst, (item: ILabelItem<number>) =>
+        intersect(item as any, firstSourceItem as any, labelGap)
+      );
+      forceItemVisible(
+        lastSourceItem,
+        items,
+        checkLast,
+        (item: ILabelItem<number>) =>
+          intersect(item as any, lastSourceItem as any, labelGap) ||
+          (checkFirst && item !== firstSourceItem ? intersect(item as any, firstSourceItem as any, labelGap) : false),
+        true
+      );
+
       const ticks = items.map(item => item.value);
 
       if (ticks.length < 3 && labelFlush) {
@@ -126,7 +197,6 @@ export const continuousTicks = (scale: ContinuousScale, op: ITickDataOpt): ITick
       scaleTicks = ticks;
     }
   }
-
   return convertDomainToTickData(scaleTicks);
 };
 
@@ -137,7 +207,7 @@ const methods = {
   greedy: function <T>(items: ILabelItem<T>[], sep: number) {
     let a: ILabelItem<T>;
     return items.filter((b, i) => {
-      if (!i || !intersect(a.AABBBounds, b.AABBBounds, sep)) {
+      if (!i || !intersect(a as any, b as any, sep)) {
         a = b;
         return true;
       }